swapped android launcher app
