+++ detox/android/detoxLauncher/app/.gitignore
