+++ detox/android/detoxLauncher/app/proguard-rules.pro
