+++ detox/android/detoxLauncher/app/src/main/res/mipmap-mdpi/ic_launcher_round.png
