+++ detox/android/detoxLauncher/app/src/main/res/mipmap-xxhdpi/ic_launcher.png
