+++ detox/android/detoxLauncher/app/src/main/res/mipmap-xxxhdpi/ic_launcher_round.png
