+++ detox/android/detoxLauncher/app/src/main/res/values/colors.xml
