+++ detox/android/detoxLauncher/app/src/main/res/values/styles.xml
