+++ detox/android/detoxLauncher/gradle.properties
