+++ detox/android/detoxLauncher/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Sun Nov 18 15:01:49 IST 2018
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.4-all.zip
