+++ detox/android/detoxLauncher/gradlew.bat
