+++ detox/android/detoxLauncher/settings.gradle
