--- detox/android/wrapperApp/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Thu Aug 30 11:28:17 IDT 2018
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-4.4-all.zip
